Isolate the two failing tests in run-not-optimised
