+++ compiler/test/dotty/tools/dotc/CompilationTests.scala
@@ -168,6 +168,10 @@ class CompilationTests extends ParallelTesting {
   @Test def runAll: Unit =
     compileFilesInDir("../tests/run", defaultOptions).checkRuns()
 
+  // The two tests that current fail under -optimise
+  @Test def runNotOptimised: Unit =
+    compileFilesInDir("../tests/run-not-optimised", defaultOptions.filterNot("-optimise".==)).checkRuns()
+
   // Pickling Tests ------------------------------------------------------------
   //
   // Pickling tests are very memory intensive and as such need to be run with a
