--- tests/pending/t7126.scala
@@ -1,11 +1,0 @@
-import language._
-
-object Test {
-  type T = Any
-  boom(???): Option[T] // SOE
-  def boom[CC[U]](t : CC[T]): Option[CC[T]] = None
-
-  // okay
-  foo(???): Option[Any]
-  def foo[CC[U]](t : CC[Any]): Option[CC[Any]] = None
-}
